[Refactor ⚙️] contexts 폴더 -> stores 폴더로 변경
